Implement basic CI: byte-compile + package-lint
The latest release of agent-shell introduced some byte-compilation errors on startup in my Emacs
environment. A snippet:

```
agent-shell-viewport.el:663:39: Warning: the function ‘agent-shell-other-buffer’ is not known to be defined.
agent-shell-viewport.el:650:48: Warning: the function ‘agent-shell-buffers’ is not known to be defined.
agent-shell-viewport.el:523:8: Warning: the function ‘agent-shell-cycle-session-mode’ is not known to be defined.
agent-shell-viewport.el:508:8: Warning: the function ‘agent-shell-set-session-mode’ is not known to be defined.
agent-shell-viewport.el:494:8: Warning: the function ‘agent-shell-set-session-model’ is not known to be defined.
agent-shell-viewport.el:488:29: Warning: the function ‘agent-shell--current-shell’ is not known to be defined.
agent-shell-viewport.el:427:28: Warning: the function ‘agent-shell--get-region’ is not known to be defined.
agent-shell-viewport.el:94:8: Warning: the function ‘agent-shell--display-buffer’ is not known to be defined.
```

To combat these in the future, this PR proposes some lightweight PR here that byte-compiles and runs
package-lint -- both in "strict" mode that treats warnings as errors.

This CI approach draws inspiration from both the lightweight CI approach taken in
purcell/package-lint [here](https://github.com/purcell/package-lint/blob/master/Makefile) and in my
own approach in jinnovation/kele.el
[here](https://github.com/jinnovation/kele.el/blob/main/.github/workflows/test.yaml).